Get only patch version updates to `pulldown-cmark`
This is for Dependabot version updates. It doesn't affect
Dependabot security updates, nor does it affect any ways of
upgrading that don't use Dependabot. See:

- https://github.blog/changelog/2021-05-21-dependabot-version-updates-can-now-ignore-major-minor-patch-releases/
- https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#ignore--

The reason is that `pulldown-cmark` will require changes to adapt
to higher versions, and for some of those changes we don't yet know
what is needed. See the full discussion in #54 and comment in #59.